API changed to Flask and asyncio task
Modified API server to run as a task under manager.

Manager can then provide the API the dictionary of game servers, and the event bus to trigger actions off.

Added a few new endpoints to interact with the API:
/api/get_instances_status now pulls states from all configured servers
/api/stop_server now stops a server with given port in the request body
/api/start_server now starts a server with given port in the request body

TOD: "Status" command is a bit ugly now. Need to fix