Workaround for game dll not found
this is an atrocious fix until I find a better solution.

on some systems, the compiled honfigurator.exe file, which is just launcher.py from cogs.misc causes issues for the opened hon_x64.exe.

The exe is unable to locate one of the game dll resources. Causing it to not open properly.

I wasted a lot of time trying to troubleshoot it, launching main.py directly works fine. Adding the hon game directory to the path is my solution until a better one comes around. It's set within the scope of the script, and doesn't modify the systems environment.